requirements.txt
torch>=2.2.1
transformers>=4.51.0
datasets>=3.5.0
sentence-transformers>=2.6.1
faiss-cpu>=1.10.0
pandas>=2.2.3
numpy>=2.2.4
tqdm>=4.66.2
nltk>=3.8.1
gradio>=5.23.3
pyarrow>=15.0.2
